nodetours/requirements.txt
# requirements.txt
anthropic>=0.15.0
langchain>=0.1.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pyyaml>=6.0.1
requests>=2.31.0
pytest>=7.4.3
openai>=1.3.0
serpapi>=0.1.0
api
requests>=2.31.0
python-dotenv>=1.0.0